+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Horse-racing duals".
 - Tests for "MIME type".
 - Tests for "Temperatures".
+- Tests for "Blunder - episode 1".
 
 ## [1.0.0] - 2022-02-26
 ### Added